requirements.txt
escapism
PyYAML
kubernetes
Jinja2
requests
click
jinja2-ansible-filters
jsonpath-ng
mkdocs